Rework of encoders
Remove the dual LED, switch and encoder footprints, it was too complex and I want encoders, I will never be using them as just mx switches.
Changed the encoder under the LCD to an EVQWGD001 since it wont block the screen and be easier to use with just your index finger.